layout/base/tests/test_resize_flush.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1279202
-->
<head>
  <title>Test for Bug 1279202</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="run()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1279202">Mozilla Bug 1279202</a>
<iframe src="resize_flush_iframe.html" id="iframe" height="200" width="200" style="border:none"></iframe>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 1279202 **/

SimpleTest.waitForExplicitFinish();

function run() {

  var iframe = document.getElementById("iframe");
  var doc = iframe.contentDocument.documentElement;
  var win = iframe.contentWindow;
  var body = iframe.contentDocument.body;

  // Flush any pending layout changes before we start.
  var width = doc.clientWidth;

  // Resize the iframe
  iframe.width = '300px';

  // Flush pending style changes, but not layout ones. We do this twice because the first flush
  // does a partial flush of the resize (setting the size on the pres context) which sets the
  // need style flush flag again. The second call makes sure mNeedStyleFlush is false.
  var color = win.getComputedStyle(body).getPropertyValue("background-color");
  color = win.getComputedStyle(body).getPropertyValue("background-color");
  is(color, "rgb(0, 128, 0)", "Style flush not completed when resizing an iframe!");

  // Query the size of the inner document and make sure it has had a layout flush.
  width = doc.clientWidth;

  is(width, 300, "Layout flush not completed when resizing an iframe!");
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
